tidyfinance: Tidy Finance Helper Functions

Helper functions for empirical research in financial economics, addressing a variety of topics covered in Scheuch, Voigt, and Weiss (2023) doi:10.1201/b23237 . The package is designed to provide shortcuts for issues extensively discussed in the book, facilitating easier application of its concepts. For more information and resources related to the book, visit https://www.tidy-finance.org/r/index.html.

See also

Author

Maintainer: Christoph Scheuch christoph@tidy-intelligence.com (ORCID) [copyright holder]

Authors:

Stefan Voigt stefan.voigt@econ.ku.dk (ORCID) [copyright holder]

Patrick Weiss patrickw@ru.is (ORCID) [copyright holder]

Other contributors:

Maximilian Mücke (ORCID) [contributor]